Refactoring?
